Added a position helper
press caps lock + 3, 4, or 5 to start a instance of this helper.
Upon initiation you click where on the screen you want to create a gui object or gui control and then you drag your mouse to where you want the bottom right corner to be. when you let go of the mouse a dialog box will pop up with the x, y, w, and h for your gui. Easy: paste and go.